Solving Bugs and Code Refactoring
Solving routing issues and refactoring code
