+++ api/index.js
@@ -4,6 +4,7 @@ const cookieParser = require("cookie-parser");
 const { default: mongoose } = require("mongoose");
 const bcrypt = require("bcryptjs");
 const User = require("./models/User.js");
+const Place = require("./models/Place.js");
 const jwt = require("jsonwebtoken");
 const imageDownloader = require("image-downloader");
 const multer = require("multer");
@@ -126,6 +127,41 @@ app.post("/upload", photosMiddleware.array("photos", 100), (req, res) => {
   }
   console.log(uploadedFiles);
   res.json(uploadedFiles);
+});
+
+// @desc Create a new place
+// route POST /places/new
+// @access Public
+app.post("/places", (req, res) => {
+  const { token } = req.cookies;
+  const {
+    title,
+    address,
+    addedPhotos,
+    description,
+    perks,
+    extraInfo,
+    checkIn,
+    checkOut,
+    maxGuests,
+  } = req.body;
+  jwt.verify(token, jwtSecret, {}, async (err, userData) => {
+    if (err) throw err;
+
+    const placeDoc = await Place.create({
+      owner: userData.id,
+      title,
+      address,
+      addedPhotos,
+      description,
+      perks,
+      extraInfo,
+      checkIn,
+      checkOut,
+      maxGuests,
+    });
+  });
+  res.json(placeDoc);
 });
 
 app.listen(4000);
