+++ client/src/pages/PlacesPage.jsx
@@ -1,160 +1,34 @@
 import { useState } from "react";
-import { Link, useParams } from "react-router-dom";
-import axios from "axios";
-import Perks from "../Perks";
-import PhotosUploader from "../PhotosUploader";
+import { Link, Navigate, useParams } from "react-router-dom";
+import AccountNav from "../AccountNav";
 
 export default function PlacesPage() {
-  // state initalizations
-  const { action } = useParams();
-  // console.log(action);
-  const [addedPhotos, setAddedPhotos] = useState([]);
-  const [title, setTitle] = useState("");
-  const [address, setAddress] = useState("");
-
-  const [description, setDescription] = useState("");
-  const [perks, setPerks] = useState([]);
-  const [extraInfo, setExtraInfo] = useState("");
-  const [checkIn, setCheckIn] = useState("");
-  const [checkOut, setCheckOut] = useState("");
-  const [maxGuests, setMaxGuests] = useState(1);
-
-  // functions
-  function inputHeader(text) {
-    return <h2 className="mt-4 text-2xl">{text}</h2>;
-  }
-
-  function inputDescription(text) {
-    return <p className="text-grey-500 text-sm">{text}</p>;
-  }
-
-  function preInput(header, description) {
-    return (
-      <>
-        {inputHeader(header)}
-        {inputDescription(description)}
-      </>
-    );
-  }
-
   return (
     <div className="m-4">
-      {action !== "new" && (
-        <div className="text-center">
-          <Link
-            className="bg-primary inline-flex gap-1 rounded-full px-6 py-2 text-white"
-            to={"/account/places/new"}
+      <AccountNav />
+      List of All Places
+      <div className="text-center">
+        <Link
+          className="bg-primary inline-flex gap-1 rounded-full px-6 py-2 text-white"
+          to={"/account/places/new"}
+        >
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            fill="none"
+            viewBox="0 0 24 24"
+            strokeWidth={1.5}
+            stroke="currentColor"
+            className="h-6 w-6"
           >
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              fill="none"
-              viewBox="0 0 24 24"
-              strokeWidth={1.5}
-              stroke="currentColor"
-              className="h-6 w-6"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                d="M12 4.5v15m7.5-7.5h-15"
-              />
-            </svg>
-            Add new Place
-          </Link>
-        </div>
-      )}
-      {action === "new" && (
-        <div>
-          <form action="">
-            {/*  */}
-            {/* Title */}
-            {preInput(
-              "Title",
-              "Title for your place. Should be short and catchy",
-            )}
-            <input
-              type="text"
-              value={title}
-              onChange={(ev) => setTitle(ev.target.value)}
-              placeholder="title, for example: My lovely apt"
+            <path
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              d="M12 4.5v15m7.5-7.5h-15"
             />
-
-            {/* Address */}
-            {preInput("Address", "Address to this place ")}
-            <input
-              type="text"
-              value={address}
-              onChange={(ev) => setAddress(ev.target.value)}
-              placeholder="Address"
-            />
-
-            {/* Photots */}
-            {preInput("Photos", " More is better")}
-
-            <PhotosUploader
-              addedPhotos={addedPhotos}
-              onChange={setAddedPhotos}
-            />
-
-            {/* Description */}
-            {preInput("Description", "Description of the place ")}
-            <textarea
-              value={description}
-              onChange={(ev) => setDescription(ev.target.value)}
-            />
-
-            {/* Perks */}
-            {preInput("Perks", "select all the perks of your place")}
-            <div className="mt-2grid-cols-2  grid gap-2 md:grid-cols-3 lg:grid-cols-6">
-              <Perks selected={perks} onChange={setPerks} />
-            </div>
-            {preInput("Extra info", "House rules ,etc")}
-            <textarea
-              value={extraInfo}
-              onChange={(ev) => setExtraInfo(ev.target.value)}
-            />
-
-            {/* Check in & check out times */}
-            {preInput(
-              "Check in&out times",
-              "Add check in and out times,remember to have some time for cleaning the room between the guests",
-            )}
-            <div className=" sm: grid grid-cols-3 gap-2">
-              <div>
-                <h3 className="-mb-1 mt-2">Check in time</h3>
-                <input
-                  type="text"
-                  value={checkIn}
-                  onChange={(ev) => setCheckIn(ev.target.value)}
-                  placeholder="14:00"
-                />
-              </div>
-              <div>
-                <h3 className="-mb-1 mt-2">Check out time</h3>
-                <input
-                  type="text"
-                  value={checkOut}
-                  onChange={(ev) => setCheckOut(ev.target.value)}
-                  placeholder="11:00"
-                />
-              </div>
-
-              {/* Max Number of guests */}
-              <div>
-                <h3 className="-mb-1 mt-2">Max number of guests</h3>
-                <input
-                  type="number"
-                  value={maxGuests}
-                  onChange={(ev) => setMaxGuests(ev.target.value)}
-                />
-              </div>
-            </div>
-            <div>
-              <button className="primary my-4">Save</button>
-            </div>
-          </form>
-        </div>
-      )}
+          </svg>
+          Add new Place
+        </Link>
+      </div>
       my places
     </div>
   );
